requirements.txt
streamlit
tensorflow
pillow
numpy
opencv-python-headless
matplotlib
requests
